+++ cli/src/main.rs
@@ -12,14 +12,14 @@
 
 #![doc = include_str!("../README.md")]
 
-/// Tool to display and create C2PA manifests
+/// Tool to display and create C2PA manifests.
 ///
 /// A file path to an asset must be provided. If only the path
 /// is given, this will generate a summary report of any claims
 /// in that file. If a manifest definition JSON file is specified,
 /// the claim will be added to any existing claims.
 use std::{
-    fs::{create_dir_all, remove_dir_all, File},
+    fs::{create_dir_all, remove_dir_all, remove_file, File},
     io::Write,
     path::{Path, PathBuf},
     str::FromStr,
@@ -195,7 +195,7 @@ struct ManifestDef {
     ingredient_paths: Option<Vec<PathBuf>>,
 }
 
-// convert certain errors to output messages
+// Convert certain errors to output messages.
 fn special_errs(e: c2pa::Error) -> anyhow::Error {
     match e {
         Error::JumbfNotFound => anyhow!("No claim found"),
@@ -206,7 +206,7 @@ fn special_errs(e: c2pa::Error) -> anyhow::Error {
     }
 }
 
-// normalize extensions so we can compare them
+// Normalize extensions so we can compare them.
 fn ext_normal(path: &Path) -> String {
     let ext = path
         .extension()
@@ -469,7 +469,7 @@ fn main() -> Result<()> {
     //     match args.output {
     //         Some(output) => {
     //             if output.exists() && !args.force {
-    //                 bail!("Output already exists, use -f/force to force write");
+    //                 bail!("Output already exists; use -f/force to force write");
     //             }
     //             if path != &output {
     //                 std::fs::copy(path, &output)?;
@@ -517,7 +517,7 @@ fn main() -> Result<()> {
         } else {
             manifest.claim_generator_info.insert(1, tool_generator);
         }
-        println!("claim generator {:?}", manifest.claim_generator_info);
+
         // set manifest base path before ingredients so ingredients can override it
         if let Some(base) = base_path.as_ref() {
             builder.base_path = Some(base.clone());
@@ -596,8 +596,12 @@ fn main() -> Result<()> {
                 if ext_normal(&output) != ext_normal(&args.path) {
                     bail!("Output type must match source type");
                 }
-                if output.exists() && !args.force {
-                    bail!("Output already exists, use -f/force to force write");
+                if output.exists() {
+                    if args.force {
+                        remove_file(&output)?;
+                    } else {
+                        bail!("Output already exists; use -f/force to force write");
+                    }
                 }
 
                 if output.file_name().is_none() {
@@ -607,10 +611,15 @@ fn main() -> Result<()> {
                     bail!("Missing extension output");
                 }
 
-                #[allow(deprecated)] // todo: remove when we can
-                builder
+                let manifest_data = builder
                     .sign_file(signer.as_ref(), &args.path, &output)
                     .context("embedding manifest")?;
+
+                if args.sidecar {
+                    let sidecar = output.with_extension("c2pa");
+                    let mut file = File::create(&sidecar)?;
+                    file.write_all(&manifest_data)?;
+                }
 
                 // generate a report on the output file
                 let reader = Reader::from_file(&output).map_err(special_errs)?;
@@ -633,7 +642,7 @@ fn main() -> Result<()> {
             if args.force {
                 remove_dir_all(&output)?;
             } else {
-                bail!("Output already exists, use -f/force to force write");
+                bail!("Output already exists; use -f/force to force write");
             }
         }
         create_dir_all(&output)?;
@@ -708,8 +717,8 @@ pub mod tests {
     #[test]
     fn test_manifest_config() {
         const SOURCE_PATH: &str = "tests/fixtures/earth_apollo17.jpg";
-        const OUTPUT_PATH: &str = "target/tmp/unit_out.jpg";
-        create_dir_all("target/tmp").expect("create_dir");
+        const OUTPUT_PATH: &str = "../target/tmp/unit_out.jpg";
+        create_dir_all("../target/tmp").expect("create_dir");
         std::fs::remove_file(OUTPUT_PATH).ok(); // remove output file if it exists
         let mut builder = Builder::from_json(CONFIG).expect("from_json");
 
@@ -719,7 +728,6 @@ pub mod tests {
             .signer()
             .expect("get_signer");
 
-        #[allow(deprecated)] // todo: remove when we can
         let _result = builder
             .sign_file(signer.as_ref(), SOURCE_PATH, OUTPUT_PATH)
             .expect("embed");